Basic draw / zoom functionality is possible, collect global statistic values for correct height adaption of statistic lines.

+++ HDJumpshot/src/base/drawable/CoordPixelXform.java
@@ -17,7 +17,7 @@ public interface CoordPixelXform
 
     public int     convertTimelineToPixel( int timeline );
     
-    public float   getTimelineHeight(int timeline);
+    public int     getTimelineHeight(int timeline);
 
     public int     convertPixelToTimeline( int vert_pixel );
